asydis

Trait CompressableDisplay

pub trait CompressableDisplay: DrawTarget {
    type BufferElement: Copy + PartialEq + Default;

    // Required methods
    fn map_to_buffer_element(color: Self::Color) -> Self::BufferElement;
    fn calculate_buffer_index(point: Point, buffer_area_size: Size) -> usize;
    async fn flush_chunk(
        &mut self,
        chunk: &[Self::BufferElement],
        chunk_area: Rectangle,
    );
}

A buffered DrawTarget that can be compressed and shared among multiple apps.

Required Associated Types

type BufferElement: Copy + PartialEq + Default

The type of elements saved to the buffer - may differ from DrawTarget::Color.

Required Methods

fn map_to_buffer_element(color: Self::Color) -> Self::BufferElement

Specify how Color maps to BufferElement.

fn calculate_buffer_index(point: Point, buffer_area_size: Size) -> usize

Calculate the buffer position of a Point.

async fn flush_chunk( &mut self, chunk: &[Self::BufferElement], chunk_area: Rectangle, )

Flushes a given chunk. Called once per chunk for every flush.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.